pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "snap-memories-tool"
version = "0.1.0"
description = "Export and process Snapchat memories: composite overlays and embed GPS metadata"
requires-python = ">=3.10"
dependencies = [
    "Pillow",
    "piexif",
    "tqdm",
]

[project.scripts]
snap-process = "snap_memories.cli:main"

[project.optional-dependencies]
dev = ["ruff", "pytest"]

[tool.setuptools.packages.find]
where = ["."]
include = ["snap_memories*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
